- Updates render methode for the merged block. - Removes standalone block files.
