--- assets/src/latest-stories-block/block/edit.css
@@ -1,35 +1,0 @@
-/* Scroll stories for carousel view. */
-.web-stories-list.is-view-type-carousel {
-  display: flex;
-  overflow-x: scroll;
-}
-
-/* Setting controls styles */
-.components-toggle-control.is-disabled .components-form-toggle__track {
-  background-color: lightgrey;
-  border-color: lightgrey;
-}
-
-.components-toggle-control.is-disabled .components-form-toggle__thumb {
-  background-color: #fff;
-  border-color: #fff;
-}
-
-.components-toggle-control.is-disabled .components-form-toggle__input,
-.components-toggle-control.is-disabled .components-toggle-control__label {
-  cursor: unset;
-}
-
-.components-input-control__input {
-  text-align: center;
-}
-
-.web-stories-settings__image-alignment .components-base-control__field {
-  display: flex;
-  align-items: center;
-}
-
-.web-stories-settings__image-alignment .components-base-control__label {
-  margin-right: 8px;
-  margin-bottom: 0;
-}
