--- assets/src/selected-stories-block/block/components/loaderContainer.js
@@ -1,51 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/**
- * External dependencies
- */
-import PropTypes from 'prop-types';
-import styled from 'styled-components';
-
-/**
- * WordPress dependencies
- */
-import { Spinner } from '@wordpress/components';
-
-const StyledContainer = styled.div`
-  display: flex;
-  height: 100%;
-  justify-content: center;
-  align-items: center;
-
-  .components-spinner {
-    margin-top: 0;
-  }
-`;
-
-function LoaderContainer({ children }) {
-  return (
-    <StyledContainer>
-      {children}
-      <Spinner />
-    </StyledContainer>
-  );
-}
-
-LoaderContainer.propTypes = {
-  children: PropTypes.node,
-};
-
-export default LoaderContainer;
