--- assets/src/selected-stories-block/index.js
@@ -1,32 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * WordPress dependencies
- */
-import { registerBlockType } from '@wordpress/blocks';
-
-/**
- * Internal dependencies
- */
-import { initializeTracking } from '../tracking';
-import { name, settings } from './block';
-
-__webpack_public_path__ = global.webStoriesSelectedBlockSettings.publicPath;
-
-registerBlockType(name, settings);
-
-initializeTracking('Selected Stories Block', false);
