+++ assets/src/web-stories-block/block/block.json
@@ -1,5 +1,5 @@
 {
-  "name": "web-stories/block",
+  "name": "web-stories/stories-list",
   "category": "embed",
   "attributes": {
     "blockType": {
