+++ includes/Block/Web_Stories_Block.php
@@ -26,14 +26,17 @@
 
 namespace Google\Web_Stories\Block;
 
-use Google\Web_Stories\Tracking;
 use Google\Web_Stories\Embed_Base;
+use Google\Web_Stories\Story_Query;
+use Google\Web_Stories\Tracking;
 use Google\Web_Stories\Story_Post_Type;
+use Google\Web_Stories\Traits\Assets;
 
 /**
  * Latest Stories block class.
  */
-class Web_Stories_Block extends Embed_Base {
+class Web_Stories_Block {
+	use Assets;
 
 	/**
 	 * Script handle.
@@ -47,7 +50,7 @@ class Web_Stories_Block extends Embed_Base {
 	 *
 	 * @var string
 	 */
-	const BLOCK_NAME = 'web-stories/block';
+	const BLOCK_NAME = 'web-stories/stories-list';
 
 	/**
 	 * Current block's block attributes.
@@ -71,16 +74,14 @@ class Web_Stories_Block extends Embed_Base {
 	 * @return void
 	 */
 	public function init() {
-		$this->register_script( self::SCRIPT_HANDLE, [ self::STORY_PLAYER_HANDLE, Tracking::SCRIPT_HANDLE ] );
-		$this->register_style( self::SCRIPT_HANDLE, [ self::STORY_PLAYER_HANDLE ] );
+		$this->register_script( self::SCRIPT_HANDLE, [ Embed_Base::STORY_PLAYER_HANDLE, Tracking::SCRIPT_HANDLE ] );
+		$this->register_style( self::SCRIPT_HANDLE, [ Embed_Base::STORY_PLAYER_HANDLE ] );
 
 		wp_localize_script(
 			self::SCRIPT_HANDLE,
 			'webStoriesBlockSettings',
 			$this->get_script_settings()
 		);
-
-		// todo: use register_block_type_from_metadata() once generally available.
 
 		// Note: does not use 'script' and 'style' args, and instead uses 'render_callback'
 		// to enqueue these assets only when needed.
@@ -216,5 +217,143 @@ class Web_Stories_Block extends Embed_Base {
 				'fields'   => [ 'ID', 'display_name' ],
 			]
 		);
+	}
+
+	/**
+	 * Initializes class variable $block_attributes.
+	 *
+	 * @since
+	 *
+	 * @param array $block_attributes Array containing block attributes.
+	 *
+	 * @return bool Whether or not block attributes have been initialized with given value.
+	 */
+	protected function initialize_block_attributes( $block_attributes = [] ) {
+		if ( ! empty( $block_attributes ) || ! is_array( $block_attributes ) ) {
+			$this->block_attributes = $block_attributes;
+			return true;
+		}
+		return false;
+	}
+
+	/**
+	 * Renders the block type output for given attributes.
+	 *
+	 * @since
+	 *
+	 * @param array $attributes Block attributes.
+	 *
+	 * @return string Rendered block type output.*
+	 */
+	public function render_block( array $attributes ) {
+
+		if ( false === $this->initialize_block_attributes( $attributes ) ) {
+			return '';
+		}
+
+		if ( ! empty( $attributes['blockType'] )
+			&& ( 'latest-stories' === $attributes['blockType'] || 'selected-stories' === $attributes['blockType'] ) ) {
+
+			$story_attributes = [
+				'align'                     => ! empty( $attributes['align'] ) ? $attributes['align'] : 'none',
+				'view_type'                 => ! empty( $attributes['viewType'] ) ? $attributes['viewType'] : 'grid',
+				'number_of_columns'         => ! empty( $attributes['numOfColumns'] ) ? $attributes['numOfColumns'] : 3,
+				'show_title'                => ! empty( $attributes['isShowingTitle'] ) ? $attributes['isShowingTitle'] : false,
+				'show_date'                 => ! empty( $attributes['isShowingDate'] ) ? $attributes['isShowingDate'] : false,
+				'show_author'               => ! empty( $attributes['isShowingAuthor'] ) ? $attributes['isShowingAuthor'] : false,
+				'show_stories_archive_link' => ! empty( $attributes['isShowingViewAll'] ) ? $attributes['isShowingViewAll'] : false,
+				'stories_archive_label'     => ! empty( $attributes['viewAllLinkLabel'] ) ? $attributes['viewAllLinkLabel'] : __( 'View all stories', 'web-stories' ),
+				'show_story_player'         => ! empty( $attributes['isShowingStoryPlayer'] ) ? $attributes['isShowingStoryPlayer'] : false,
+				'list_view_image_alignment' => ! empty( $attributes['imageOnRight'] ) ? 'right' : 'left',
+				'has_square_corners'        => ! empty( $attributes['isStyleSquared'] ) ? $attributes['isStyleSquared'] : false,
+			];
+
+			$stories = new Story_Query( $story_attributes, $this->get_query_args() );
+
+			return $stories->render();
+		}
+
+		$embed_block = new Embed_Base();
+
+		return $embed_block->render( $attributes );
+	}
+
+	/**
+	 * Returns arguments to be passed to the WP_Query object initialization.
+	 *
+	 * @since
+	 *
+	 * @param array $attributes Current block's attributes. If not passed, will use attributes stored in class variable.
+	 *
+	 * @return array Query arguments.
+	 */
+	protected function get_query_args( array $attributes = [] ) {
+
+		if ( empty( $attributes ) ) {
+			$attributes = $this->block_attributes;
+		}
+
+		if ( empty( $attributes ) ) {
+			return [];
+		}
+
+		$query_args = [
+			'post_type'        => Story_Post_Type::POST_TYPE_SLUG,
+			'post_status'      => 'publish',
+			'suppress_filters' => false,
+			'no_found_rows'    => true,
+		];
+
+		// if block type is 'selected-webstories'.
+		if ( ! empty( $attributes['blockType'] )
+			&& 'selected-stories' === $attributes['blockType']
+		) {
+			// if no stories are selected return empty array.
+			if ( empty( $attributes['stories'] ) ) {
+				return [];
+			}
+
+			$query_args['post__in'] = $attributes['stories'];
+			$query_args['orderby']  = 'post__in';
+
+			return $query_args;
+		}
+
+		$order_by_value = ( ! empty( $attributes['orderByValue'] ) ) ? $attributes['orderByValue'] : '';
+		$num_of_stories = ( ! empty( $attributes['numOfStories'] ) ) ? absint( $attributes['numOfStories'] ) : '';
+
+		if ( ! empty( $num_of_stories ) ) {
+			$query_args['posts_per_page'] = $num_of_stories;
+		}
+
+		if ( ! empty( $order_by_value ) ) {
+			switch ( $order_by_value ) {
+				case 'old-to-new':
+					$query_args['order'] = 'ASC';
+					break;
+				case 'alphabetical':
+					$query_args['orderby'] = 'title';
+					$query_args['order']   = 'ASC';
+					break;
+				case 'reverse-alphabetical':
+					$query_args['orderby'] = 'title';
+					$query_args['order']   = 'DESC';
+					break;
+				case 'random':
+					$query_args['orderby'] = 'rand';
+					$query_args['order']   = 'DESC';
+					break;
+			}
+		}
+
+		if ( ! empty( $attributes['authors'] ) && is_array( $attributes['authors'] ) ) {
+			$author_ids = wp_list_pluck( $attributes['authors'], 'id' );
+
+			if ( ! empty( $author_ids ) && is_array( $author_ids ) ) {
+				$query_args['author__in'] = $author_ids;
+			}
+		}
+
+		return $query_args;
 	}
 }
