+++ includes/Plugin.php
@@ -41,8 +41,6 @@ use Google\Web_Stories\Block\Embed_Block;
 use Google\Web_Stories\REST_API\Stories_Settings_Controller;
 use Google\Web_Stories\REST_API\Stories_Users_Controller;
 use Google\Web_Stories\Shortcode\Embed_Shortcode;
-use Google\Web_Stories\Block\Latest_Stories_Block;
-use Google\Web_Stories\Block\Selected_Stories_Block;
 use Google\Web_Stories\Block\Web_Stories_Block;
 
 /**
@@ -97,20 +95,6 @@ class Plugin {
 	 * @var Embed_Block
 	 */
 	public $embed_block;
-
-	/**
-	 * Latest Stories Block.
-	 *
-	 * @var Latest_Stories_Block
-	 */
-	public $latest_stories_block;
-
-	/**
-	 * Latest Stories Block.
-	 *
-	 * @var Selected_Stories_Block
-	 */
-	public $selected_stories_block;
 
 	/**
 	 * Web Stories Block.
@@ -224,13 +208,9 @@ class Plugin {
 		add_action( 'init', [ $this->embed_base, 'init' ], 9 );
 
 		// Gutenberg Blocks.
-		$this->embed_block            = new Embed_Block();
-		$this->latest_stories_block   = new Latest_Stories_Block();
-		$this->selected_stories_block = new Selected_Stories_Block();
-		$this->web_stories_block      = new Web_Stories_Block();
+		$this->embed_block       = new Embed_Block();
+		$this->web_stories_block = new Web_Stories_Block();
 		add_action( 'init', [ $this->embed_block, 'init' ] );
-		add_action( 'init', [ $this->latest_stories_block, 'init' ] );
-		add_action( 'init', [ $this->selected_stories_block, 'init' ] );
 		add_action( 'init', [ $this->web_stories_block, 'init' ] );
 
 		// Embed shortcode.
